% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKS_labels.R
\docType{data}
\name{SKS_labels}
\alias{SKS_labels}
\title{Data from the SKS dataset}
\format{
A data frame with ? rows and ? variables:
\describe{
  \item{Prefix}{Prefix kode kategori}
  \item{Kode}{SKS kode}
  \item{Label}{Label på dansk}
}
}
\source{
\href{https://filer.sundhedsdata.dk/sks/data/skscomplete/}{Website}
}
\usage{
data(SKS_labels)
}
\description{
Data from the SKS dataset
}
\examples{
data(SKS_labels)


}
\keyword{datasets}
