% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOMED_labels.R
\docType{data}
\name{SNOMED_labels}
\alias{SNOMED_labels}
\title{Data from the SKS dataset}
\format{
A data frame with ? rows and ? variables:
\describe{
  \item{Kode}{SNOMED kode}
  \item{Label}{Label på dansk}
}
}
\source{
\href{https://www.patobank.dk/snomed/}{Website}
}
\usage{
data(SNOMED_labels)
}
\description{
Data from the SKS dataset
}
\examples{
data(SNOMED_labels)


}
\keyword{datasets}
